*Music  +Added Music in-game.  +Cleaned the music code and make it easier to use in the future
 -Will play with the volume if needed
